Fix up the types
