+++ Composer/packages/server/src/models/settings/defaultSettingManager.ts
@@ -101,7 +101,6 @@ export class DefaultSettingManager extends FileSettingManager {
         hostname: '',
         qnaRegion: 'westus',
       },
-      orchestrator: {},
       telemetry: {
         logPersonalInformation: false,
         logActivities: true,
