+++ Composer/packages/types/src/orchestrator.ts
@@ -1,8 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+export type ModelTypes = 'en_intent' | 'multilingual_intent';
+
 export type OrchestratorModelRequest = {
-  kind: 'en_intent' | 'multilingual_intent';
+  kind: ModelTypes;
   name: string;
 };
 
@@ -10,4 +12,18 @@ export enum DownloadState {
   STOPPED = 'STOPPED',
   ALREADYDOWNLOADED = 'ALREADYDOWNLOADED',
   DOWNLOADING = 'DOWNLOADING',
+}
+
+export interface IOrchestratorNLRList {
+  version: string;
+  defaults: Record<ModelTypes, string>;
+  readonly models: Record<
+    string,
+    {
+      releaseDate: string;
+      modelUri: string;
+      description: string;
+      minSDKVersion: string;
+    }
+  >;
 }
